app/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
Flask-RESTful==0.3.8
SQLAlchemy==1.3.19
PyMySQL==0.10.0
requests-oauthlib==1.3.0
pyjwt==1.4.2
python-dotenv==0.14.0
mecab-python3==1.0.1